[VACE-2649] Add manifest.json and i18n.json files to plugin seed 9.7 - 10.0 in @vcd/schematics
Add assets for 9.7 - 10.1 plugins generated by the seed schematic.
Make the seed templates more dynamic, i.e the user will be able to set:
- manifest name
- vendor
- urn
- module name
- module file name
- containing folder name
- update angular json
- update the root plugin registry file
- add the plugin name to i18n.json file.
- ...

Testing Done:
- plugin seed works
- assets are in place
- every dynamic string is replaced with its value
