+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/schema.d.ts
@@ -1,4 +1,6 @@
 export interface Schema {
     name: string;
     vcdVersion: string;
+    module: string;
+    vendor: string;
 }
