+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/9.1/__name@dasherize__/src/main/__module@dasherize__.module.ts
@@ -1,0 +1,40 @@
+import { CommonModule } from "@angular/common";
+import { Inject, NgModule } from "@angular/core";
+import { RouterModule, Routes } from "@angular/router";
+import { Store } from "@ngrx/store";
+import { VcdApiClient, VcdSdkModule } from "@vcd/sdk";
+import { ExtensionNavRegistration, EXTENSION_ROUTE } from "@vcd/sdk/common";
+import { PluginModule } from "@vcd/sdk/core";
+import { TranslateService } from "@vcd/sdk/i18n";
+import { ClarityModule } from "clarity-angular";
+import { SimpleComponent } from "./simple/simple.component";
+
+const ROUTES: Routes = [
+    { path: "", component: SimpleComponent }
+];
+
+@NgModule({
+    imports: [
+        ClarityModule,
+        CommonModule,
+        VcdSdkModule,
+        RouterModule.forChild(ROUTES)
+    ],
+    declarations: [
+        SimpleComponent
+    ],
+    bootstrap: [SimpleComponent],
+    exports: [],
+    providers: [VcdApiClient]
+})
+export class <%= classify(module) %>Module extends PluginModule {
+    constructor(appStore: Store<any>, @Inject(EXTENSION_ROUTE) extensionRoute: string, translate: TranslateService) {
+        super(appStore, translate);
+        this.registerExtension(<ExtensionNavRegistration>{
+            path: extensionRoute,
+            icon: "page",
+            nameCode: "nav.label",
+            descriptionCode: "nav.description"
+        });
+    }
+}
