+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/9.1/__name@dasherize__/src/main/index.ts
@@ -1,2 +1,2 @@
-export {SimplePluginModule} from "./simple-plugin.module";
+export {<%= classify(module) %>Module} from "./<%= dasherize(module) %>.module";
 export {SubnavPluginModule} from "./subnav-plugin.module";
