+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/__module@dasherize__.module.ts
@@ -2,11 +2,9 @@ import { CommonModule } from "@angular/common";
 import { Inject, NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 import { Store } from "@ngrx/store";
-import { VcdApiClient, VcdSdkModule } from "@vcd/sdk";
-import { ExtensionNavRegistration, EXTENSION_ROUTE } from "@vcd/sdk/common";
-import { PluginModule } from "@vcd/sdk/core";
-import { TranslateService } from "@vcd/sdk/i18n";
-import { ClarityModule } from "clarity-angular";
+import { VcdApiClient, VcdSdkModule, PluginModule, ExtensionNavRegistration, EXTENSION_ROUTE, EXTENSION_ASSET_URL } from "@vcd/sdk";
+import { TranslationService, I18nModule} from "@vcd/i18n";
+import { ClarityModule } from "@clr/angular";
 import { SimpleComponent } from "./simple/simple.component";
 
 const ROUTES: Routes = [
@@ -18,6 +16,7 @@ const ROUTES: Routes = [
         ClarityModule,
         CommonModule,
         VcdSdkModule,
+        I18nModule.forChild(EXTENSION_ASSET_URL, true),
         RouterModule.forChild(ROUTES)
     ],
     declarations: [
@@ -27,9 +26,10 @@ const ROUTES: Routes = [
     exports: [],
     providers: [VcdApiClient]
 })
-export class SimplePluginModule extends PluginModule {
-    constructor(appStore: Store<any>, @Inject(EXTENSION_ROUTE) extensionRoute: string, translate: TranslateService) {
-        super(appStore, translate);
+export class <%= classify(module) %>Module extends PluginModule {
+    constructor(appStore: Store<any>, @Inject(EXTENSION_ROUTE) extensionRoute: string, translate: TranslationService) {
+        super(appStore);
+        translate.registerTranslations(null);
         this.registerExtension(<ExtensionNavRegistration>{
             path: extensionRoute,
             icon: "page",
