+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/simple/simple.component.html
@@ -1,0 +1,6 @@
+<div>
+  <p><i>Logged into {{tenant | async}} as {{username | async}}</i></p>
+  <h1>{{"tos.title" | translate}}</h1>
+  <p>{{"tos.1" | translate}}</p>
+  <p>{{"tos.2" | translate}}</p>
+</div>
