--- typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/subnav/about.component.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright 2018 VMware, Inc. All rights reserved. VMware Confidential
- */
-import {Component, Inject} from "@angular/core";
-import {EXTENSION_ASSET_URL} from '@vcd/sdk';
-
-@Component({
-    selector: "vcd-plugin-about",
-    template: `
-        <div>
-            <p>{{"subnav.about.content.1" | translate}}</p>
-            <p>{{"subnav.about.content.2" | translate}}</p>
-        </div>
-
-    `
-})
-export class AboutComponent {
-    constructor(@Inject(EXTENSION_ASSET_URL) public assetUrl: string) {}
-}
