--- typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/subnav/status.component.ts
@@ -1,20 +1,0 @@
-/*
- * Copyright 2018 VMware, Inc. All rights reserved. VMware Confidential
- */
-import {Component, Inject} from "@angular/core";
-import {EXTENSION_ASSET_URL} from '@vcd/sdk';
-
-@Component({
-    selector: "vcd-plugin-status",
-    template: `
-        <div>
-            <h1>{{"subnav.status.content" | translate:today}}</h1>
-        </div>
-
-    `
-})
-export class StatusComponent {
-    today: String = new Date().toLocaleString();
-    
-    constructor(@Inject(EXTENSION_ASSET_URL) public assetUrl: string) {}
-}
