--- typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/subnav/subnav.component.html
@@ -1,9 +1,0 @@
-<div class="content-area">
-  <router-outlet></router-outlet>
-</div>
-<clr-vertical-nav [clrVerticalNavCollapsible]="true">
-  <a *ngFor="let navItem of navItems" clrVerticalNavLink routerLink="{{navItem.routerLink}}" routerLinkActive="active">
-    <clr-icon *ngIf="navItem.iconShape" clrVerticalNavIcon [attr.shape]="navItem.iconShape" title="{{navItem.labelKey | translate}}"></clr-icon>
-    {{navItem.labelKey | translate}}
-  </a>
-</clr-vertical-nav>
