--- typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/main/subnav/subnav.component.ts
@@ -1,16 +1,0 @@
-import {Component, Inject} from "@angular/core";
-import {EXTENSION_ASSET_URL} from '@vcd/sdk';
-
-@Component({
-    selector: "plugin-subnav",
-    templateUrl: "./subnav.component.html",
-    host: {'class': 'content-container'}
-})
-export class SubnavComponent {
-    navItems: any[] = [
-        {routerLink: "./status", iconShape: "help-info", labelKey: "subnav.menu.status"},
-        {routerLink: "./about", iconShape: "helix", labelKey: "subnav.menu.about"}
-    ];
-    
-    constructor(@Inject(EXTENSION_ASSET_URL) public assetUrl: string) {}
-}
