+++ typescript/api-client/projects/vcd/schematics/src/schematics/plugin-seed/templates/wellington-xendi/__name@dasherize__/src/public/i18n.json
@@ -4,11 +4,6 @@
         "tos.1": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Tempus urna et pharetra pharetra massa massa. Ipsum consequat nisl vel pretium. Sagittis id consectetur purus ut faucibus pulvinar elementum integer enim. Nulla pellentesque dignissim enim sit amet. Ultrices in iaculis nunc sed augue. Tortor vitae purus faucibus ornare suspendisse sed nisi lacus. Sagittis eu volutpat odio facilisis mauris sit amet massa. Dolor magna eget est lorem ipsum. Eu mi bibendum neque egestas congue. Mi ipsum faucibus vitae aliquet nec ullamcorper sit amet. Quisque sagittis purus sit amet. Sem nulla pharetra diam sit amet nisl.",
         "tos.2": "Faucibus et molestie ac feugiat sed lectus vestibulum mattis. Rhoncus urna neque viverra justo nec ultrices dui. Gravida dictum fusce ut placerat. Facilisis gravida neque convallis a cras semper. Quis eleifend quam adipiscing vitae proin sagittis nisl rhoncus. Vitae auctor eu augue ut lectus arcu bibendum at. Sociis natoque penatibus et magnis. Eget lorem dolor sed viverra. Non curabitur gravida arcu ac. Viverra maecenas accumsan lacus vel facilisis volutpat est. Et malesuada fames ac turpis egestas sed tempus urna et. Viverra ipsum nunc aliquet bibendum enim. Nisl rhoncus mattis rhoncus urna neque. Risus in hendrerit gravida rutrum quisque non tellus orci. Orci eu lobortis elementum nibh. Eget mi proin sed libero enim sed faucibus turpis in.",
         "nav.label": "<%= classify(name) %>",
-        "nav.description": "A basic plugin project",
-        "subnav.menu.status": "Status",
-        "subnav.menu.about": "About",
-        "subnav.status.content": "Current time is {0} and the plugin status is WORKING",
-        "subnav.about.content.1": "This is an example plugin that showcases sub-navigation, globalization, and styling",
-        "subnav.about.content.2": "To build your own plugin, clone the seed project at https://github.com/vmware/vcd-ext-sdk"
+        "nav.description": "A basic plugin project"
     }
 }
